Fix exit keyword in symbol-highlight TS
problem:
The symbol highlight exit function is never called,
after the keyword: `:before-exit`
in the definition of the symbol-highlight transient state (TS).

cause:
The TS definition gets the exit function with the keyword: `:on-exit`

Then it assigns that exit function to the hydra with the keyword: `:before-exit`

That's where the confusion came from.