[defaults] Disable line numbers with `SPC t n n`
Toggles are great and all, but the trouble with the various toggles for line
numbers is that they force you to remember and think about the fact that there
are variant forms of line numbers. To beginners or forgetful people the various
combinations of behavior can be unintuitive here (e.g. toggling off relative
line numbers does nothing if you are actual in visual line numbers mode). Though
it doesn't fit exactly into the normal stuff we have for toggles, due to the
complexity of this situation I think it is good to just have a way to say "line
numbers be gone!"

Wasn't sure what keybinding to use, so used a double-tap mnemonic.

If this PR gets rejected, there are a couple minor fixes that should probably be
pulled out and submitted separately.