[compleseus] Fix keys
problem
- `SPC h d a` is undefined
- `SPC ?` shows: SPC ?-  which-key: There are no keys to show
- `SPC f b` calls: `bookmark-jump`. Setting a bookmark, requires typing the
bookmark name and calling an embark action:
 M-o                    ;; embark-act
 s                      ;; bookmark-set
- `SPC f f` is bound to `find-file` but it calls `ido-find-file`

solution
- `SPC h d a` `consult-apropos`
- `SPC ?` `embark-bindings`
- `SPC f b` `consult-bookmark`, it allows for jumping to a bookmark
(just like `bookmark-jump`), but a bookmark can be set by just typing a new
name. If there is a bookmark with a similar name, then the selection can be
moved up to the prompt. (the embark action method also works here)
- `SPC f f` `spacemacs/compleseus-find-file` (it calls `find-file`)