[major-modes] add evil-matchit support for matlab
If `evil-matchit` is available, configure it for us in `matlab-mode`.

Note: This commit by itself doesn't actually do anything, because currently,
`evil-matchit` supports `octave-mode`, but not `matlab-mode`, although it's a
very simple matter of just turning on its `octave-mode` configuration for
`matlab-mode`. I'm submitting a PR against `evil-matchit` in order to support
`matlab-mode` out of the box.