latex layer add auto configure pdf-tools and open in split window
This commit makes the latex layer automatically configure pdf-tools as
latex pdf-viewer if the pdf layer is installed.

Additionally it adds a layer variable to configure pdf-tools to open in a
split-window. (Strangely, finding out how to achieve that is far from trivial,
and also not easy to google, because many hits do not provide the "correct"
answer (see the typical example
[here](https://emacs.stackexchange.com/questions/28056/spacemacs-latex-how-to-make-view-command-tile-vertically-with-emacs-window)).