Fix keybindings in ivy-occur-grep-mode (issue #15219)
Ivy occur grep mode is a read-only mode that is very similar to ivy-occur mode,
and has one extra keybinding under `w` to change to `wgrep-mode`.

Therefore, instead of starting it normal mode, it should start with the
`overriding-map` similar as how it is done for ivy-occur-mode.

Theoretically (this currently does not work here), this should make the `w` call
`spacemacs/ivy-wgrep-change-to-wgrep-mode` which is perfectly fine.