help/helpful: Set tab-width to 8, better integration w/ Ivy
Currently we set default tab-width to 2, but in Emacs's Elisp source code,
tab-width defaults to 8.

This discrepancy results in bad alignment of Elisp source code in help-mode
and helpful-mode (when helpful layer is enabled).

This commit fixes this problem by setting tab-width to 8 in those buffers.

In addition, when both helpful and ivy are in use, this commit improved
their integration according to helpful's documentation.