Fix #15255 multi-line byte-compile warnings
The package is conditionally frozen to an earlier version if emacs-version is not >= 29.
Fix version comparison in #15255 fix.

The objects were swapped, and now the predicate will work better.
Use version<

Changes the predicate to use Spacemacs convention for version comparison.
Use a string for version.

Fix usage of `version<`.