spacemacs-buffer: lock it in read-only
This commit locks the spacemacs home buffer in read-only mode.

Before this, users may use `SPC b w` to toggle it to writable but
this buffer really should not be modified by users.

Note that, when we need to modify the home buffer in LISP program,
we can always set `inihibit-read-only` to nil. Thus this commit
won't disallow us to update the home buffer when it's needed.