moved `all-the-icons` from `auto-completion` to `spacemacs-visual`
This change is due to the fact that `all-the-icons` is used in
`core/spacemacs-buffer.el`, which doesn't belong to any layer.

On a new installation of Spacemacs, `all-the-icons` is not installed yet but
also note that `spacemacs-buffer-show-icons` defaults to nil, thus no error will
occur.

Once the user modified the said variable to t, `all-the-icons` would be
installed.