core-jump: fixed a typo
For example, in `rust-mode`, `(spacemacs//get-jump-handlers)` returns
`((lsp-ui-peek-find-definitions :async t) evil-goto-definition)`.

So `-handler` is bound to the first one, i.e. `(lsp-ui-peek-find-definitions :async t)`.
We need to figure out whether it's asynchronous by checking the `:async` keyword argument.

But before this commit, we wrongly checks the second handler in the handlers list.
Even when multiple handlers do exist and they are all async, the typo made it sub-optimal since it uses not the best handler.

This commit fixes that.