Add info about auth-sources for configuring SSH keyrings
co-authored-by: bryce-carson bcars268@mtroyal.ca

This commit adds a sub-section to the ~README.org~ describing the configuration
of the auth-sources variable, which enables Emacs (and Magit) to access Login or
session keyrings managed by a Linux desktop environment.

This prevents the user from needing to write a ~.gitconfig~ or a ~.netrc~ for
storing Git forge authentication tokens hazardously, or with PGP (GPG?).