[helm] fix typo in helm-ls-git
helm-ls-git is specified as a package belonging to the helm layer, but it's
declared with the (non-existent) keyword :require, which should perhaps be
:requires. Changing it to :requires means that helm-ls-git will be installed
only if the git *package* is installed. It seems more likely that the original author's
intent was to install helm-ls-git only if the git *layer* was installed.

So, now we use a :toggle (h/t @lebensterben) so that `helm-ls-git` is toggled on
if the git layer is installed, but not otherwise.