[core] Make startup not fail when home buffer is customised
in #15421 core cannot find a certain text in the home buffer
to start adding additional elements. This did fail the
start up process.

Now if the starting point is not found the cursor is set
to the end of the buffer instead and no error is signaled.

This allows to still have the home buffer work as expected
even when the starting point has been somehow removed.