[helm] Fix dangerous implementation for helm-use-fuzzy
This variable caused Spacemacs to set all helm sources
to use fuzzy matching independently of whether this
was supported by the source.

This handling was very dangerous as using fuzzy for an
incompatible source can have all kind of side effects
causing unique bugs which will never be fixed by the
source package authors.

Therefore I have changed the implementation to
activate fuzzy matching via helm variables instead,
which should make sure we only activate this for
compatible sources.

If sources are found which do not longer do
fuzzy matching, the individual source should be
improved to support this feature and declare it's
compatibility towards helm.

We cannot fix helm in Spacemacs, instead the respective
package must be patched. Everything else is a maintenance
nightmare!

Fixes #15124