[git] Disable evil-surround-mode in all magit modes
This is an extension of #15462.  I noticed that the fix did not apply
to magit-diff buffers, which still had the issue of `s' not being
available to stage diff hunks.  `magit-diff-mode' is the parent mode
of all magit modes, in which evil-surround is not useful, so disabling
there makes the most sense.