Fix a previous attempted fix loading mu4e.
mu version 1.8.2 renamed mu-meta to mu-config.  As the variable mu4e-mu-version
is defined in said file the easiest fix seems to be to simply attempt to load
the new one, and only if that fails to load the old version.

Maybe this can be reworked or better yet removed in the future.