[spacemacs-bootstrap] Remove whitespace from kill ring.
dotspacemacs-enable-paste-transient-state is in our dotfile template, and I find
it very useful. There exists a minor annoyance that deleted whitespace with the
dd or x keys appear in the kill ring. This requires an extra C-j/C-k to bypass
the noise. Perhaps others will find it convenient to default to ignoring
whitespace in the kill-ring.

Credit: https://stackoverflow.com/questions/12102554/emacs-skip-whitespace-kills