spacemacs-editing-visual: enable term-cursor
`term-cursor` was added to provide configurable cursor shape and colour when
Emacs is running in a terminal. Originally, it would not be enabled when the
user is using `evil` editing style, as `evil-terminal-cursor-changer` seems to
be a better alternative.

As explained in commit #d48bc8d9 `evil-terminal-cursor-changer` has buggy
behaviours and was removed. This commit therefore enables `term-cursor` for all
editing styles in terminal Emacs.