Switch to maintained fork of evil-escape
this includes most of the non merged PRs
from the original repo. Right now it is
pulled from Github directly as the
original package is still unmaintained
and may be removed from melpa again.

If the package is already installed
the old package must be removed from
your elpa directory manually to
trigger the reinstallation.