Deletes uses of org-agenda-filter-by-tag-refine, as Org 9.0 removed it
The Org Mode 9.0 release removed the org-agenda-filter-by-tag-refine
and recommends the use of org-agenda-filter-by-tag in it's place. This
commit removes all uses of the function and the mentions of it in the
documentation. The same functionality can be achieved by using the typing
prefix C-u twice and then using org-agenda-filter-by-tag ("ft") as the
Org Mode documentation states.

Relevant Links:
- https://orgmode.org/worg/org-release-notes.html#org268651e
- https://orgmode.org/org.html#Filtering_002flimiting-agenda-items