Remove forced overwrite for whitespace
For literal years now I've been thinking there's something wrong with my
themes, because whitespace-mode would use the wrong colour for
highlighting spaces. Turns out it was this piece of config that was
messing up the themes.

If changes like this are needed, then they should be imlpemented by
theme maintainers, not us.