[command-log] update keybindings & modeline insert after
- Add keybindings for header and tab bar placement of keycast
- Update command-log documentation
- Update keycast-insert-after to keycast-mode-line-insert-after

tab bar provides places the keycast output in the top right of the frame and
output does not hide when using menu bar like other modes dot

header position is show on the top left of the frame (under the tab bar if
visible). header keycast output is not visible when using which-key menu or in
buffers such as commit_editmsg

keycast modeline replaces the whole modeline with its own output which is not the desired experience.
Ideally Keycast should be shown along with other important aspects of the
modeline (screenshot in issue #14169

Resolve: #14169