When user setting is nil, don't init loading progress bar
When the user customizes `dotspacemacs-loading-progress-bar` and sets it to nil,
do not initialize the loading progress bar during startup.

This prevents the full progress bar from overtaking the mode-line, preventing
redisplays that are unnecessary when no progress bar updates will occur later
for visual fluff.

In short, don't show an empty progress bar if we don't intend to fill it.