core-fonts-support: Don't fiddle with user's face Customization
Fixes #5353.

Related:
https://github.com/emacs-mirror/emacs/commit/eae23d60f8338ea4e8617b13f4f6aa06333f68cd
https://emacs.stackexchange.com/questions/33403/customize-creates-custom-set-faces-unintentionally

An alternative strategy is to call `(set-frame-font fontspec nil nil)`.
This will work on any emacs version, and this will also inhibit
fiddling with the Customization. But this will apply the font setting
only to the current frame. While this *should* be okay (because when this
code runs, only the current frame should be graphical and thus relevant
when it comes to font changes), I'm not convinced that this analysis is
correct, and that it's a good idea to rely on this fact.