Fix: Void symbol helm-source-comint-input-ring (#16148)
* [shell] Fix: Void symbol helm-source-comint-input-ring

helm-source-comint-input-ring was a symbol from from helm-comint.el, and
helm-comint was separated from helm-core as of commit

https://github.com/emacs-helm/helm/commit/587bed0f735c7955ab385a5fbc2660a61713eb29

See also: https://github.com/emacs-helm/helm/issues/2615

The package has been submitted to MELPA but helm-source-comint-input-ring has
been renamed to helm-comint-input-ring to satisfy Elisp package naming
conventions to qualify for submission onto MELPA.

The comint ring history is still bound to ", H". It is the author's opinion that
it remains immensely useful and takes advantage of the built-in Emacs comint.el
and any Emacs modes derived from it.

* [shell] Fix: initialize helm-comint in spacemacs.

    Warnings: (w)
        - package helm-comint not initialized in layer shell, you may consider removing
          this package from the package list or use the :toggle keyword instead of a
          `when' form.