[rust] Resurrect rust-backend variable
rust-backend was removed but is still meaningful (needs to be set to
the symbol lsp in order for LSP support to be enabled).

This commit restores the variable to config.el and removes boundp
checks because it will always be bound now (except in layers.el which
is loaded before config.el).