[rust] Remove wrapper commands for package cargo
This package was removed in the Rust layer overhaul, so these commands
are not used anywhere.  In fact, the reloading behavior cannot easily
be implemented with rustic, because rustic runs `cargo add` and the
like in a compilation-mode buffer rather than synchronously.

As a result, we also remove references to automatically reloading the
workspace in rust-analyzer and the corresponding config variable.