Update/improve prefix functionality in window delete commands
(Spac)Emacs comes with the `kill-buffer-and-window` by default. Also deleting
other windows (and buffers) simply by prefixing the window delete commands with
a numerical prefix is very convenient. Therefore, this commit updates (or adds)
the functionality of the prefix in the window delete commands.

Summary: to delete another window (and optionally its buffer), simply prefix
`SPC w d` (or `SPC w x`) with the window number.