load dired-x after loading dired other than use autoloading
dired-x used to be autoloaded with command dired-jump, which is the major
entrance to dired in spacemacs. Now dired-jump was moved to dired.el so this
autoloading is not needed for latest dired. However dired-x still provides
additional key bindings such as "* ." to dired. To keep the old behavior, load
dired-x after dired.

For users with older dired that dired-jump is in dired-x, we still need the auto
loading to make sure dired-jump got loaded when we call it. That is because most
of the dired functions have autoload and don't need manual specifying but
dired-jump was defined in dired-x and got autoloaded only when dired.el got
loaded, so when spacemacs provides "SPC f d" to dired-jump, at that moment,
dired.el was not loaded yet, we need to make sure dired-x got loaded when we
call.