[helpful] Remove redundant key bindings
Now that the helpful layer just aliases describe-function and so on,
there's no need to make the key bindings explicitly refer to the
helpful.el versions of those functions.  Similarly for counsel.

In fact, keeping the original key bindings means that, with the ivy
layer, SPC h d f uses counsel-describe-function instead, which is a
better experience than helpful-callable directly (the former has a
proper minibuffer history variable and includes more symbols).