Remove `''` from pdf-outline-buffer-mode-map and update documentation (#16316)
* Remove `''` from pdf-outline-keymap and update documentation

I noticed some noise in my *Messages* buffer about keybinding whenever I tried
to view a PDF, and I tracked it down.

`pdf-outline-minor-mode` gets evilified. Prior to this, `SPC` is mapped to
`pdf-outline-display-link'. During evilification, this gets remapped to
`'` (this remapping is, AFAICT, not documented in evilification but you can see
it
[here](https://github.com/syl20bnr/spacemacs/blob/4a227fc94651136a8de54bcafa7d22abe1fa0295/layers/%2Bdistributions/spacemacs-bootstrap/local/evil-evilified-state/evil-evilified-state.el#L323) (note:
`32` means `SPC`).

We then try to map the key `''` to `pdf-outline-move-to-current-page`, but this
isn't possible, because `'` is already mapped to a function and is thus not a
prefix key.

Fortunately, this behavior is already available via double-backtick.

I've removed the (failed) mapping of `''` and also removed it frmo the documentation.

* document additional keys moved in evilified state