[spacemacs-defaults] Do not prompt to revert from auto-save file
SPC b R passes nil to the IGNORE-AUTO argument of revert-buffer, which
causes it to prompt the user whether they want to revert from the
auto-save file.  This is not the default behavior of revert-buffer;
somewhat confusingly, IGNORE-AUTO is t by default when revert-buffer
is called interactively.

Changing Spacemacs to use the default behavior of revert-buffer makes
more sense here, as reverting from the auto-save file is rarely
useful.