[compleseus] Initial input for project search in symbol-highlight ts
Unlike the 's' and 'f' bindings in the Symbol Highlight Transient State, the
project search binding '/' currently does not automatically insert the symbol at
point into the prompt. This commit switches in the correct function
`spacemacs/compleseus-search-projectile`. (The only difference between
`spacemacs/compleseus-search-projectile-auto` and
`spacemacs/compleseus-search-projectile` is the initial input.)