[symbol-highlight-ts] Do not restart the transient state unnecessarily
Previously the ts was restarted on every invocation of
`spacemacs//quick-ahs-move'. This is not necessary and led to the
:on-exit-function being called every time.

Instead, we now only start the transient state in all of the entry points:
`spacemacs/enter-ahs-forward', `spacemacs/enter-ahs-backward',
`spacemacs/goto-last-searched-ahs-symbol' and `spacemacs/symbol-highlight'.

Note that `spacemacs//ahs-setup' needs to be called before every entry
point. In `spacemacs//quick-ahs-move' we still have to call `ahs-highlight-now'
to prevent blocking.

Finally, the change allows us to simplify
`spacemacs//disable-symbol-highlight-after-ahs-ts-exit'.