[go] Disable use of company-go when gopls is used
The default backend for go-layer is lsp but company-go was still
added to company-backends forcing it to search for gocode which should
not be used since the user wants lsp.

This change only adds company-go to company-backends if go-mode is
selected as the backend.

Note: go-mode is deprecated and gocode is not maintained anymore